관리자 페이지 - 매출 상세내역 ajax로 수정, pr 내용 반영하여 수정
