+++ src/main/java/com/bcc/web/AdminController.java
@@ -22,27 +22,6 @@ public class AdminController {
 	@Inject
 	private AdminService service;
 	
-	// '전체회원 보기' 클릭시 날짜에 해당하는 모든 프리미엄 회원 출력
-	@RequestMapping(value = "/pmMembers", method = RequestMethod.GET)
-	public String pmMemberGET(@RequestParam("date") String date, Model model) {
-		model.addAttribute("pmMems", service.getPmInfo(date));
-		
-		model.addAttribute("totalMem", service.getTotalMem());
-		model.addAttribute("totalPmMem", service.getTotalPmMem());
-		model.addAttribute("totalComment", service.getTotalComment());
-		model.addAttribute("totalAcc", service.getTotalAcc());
-		
-		return "/admin/profitChart";
-	}
-	
-	// '구독회원만 보기' 클릭시 날짜에 해당하는 구독회원 정보만 출력
-	@RequestMapping(value = "/periodMems", method = RequestMethod.GET)
-	public String onlyPeriodicGET(@RequestParam("date") String date, Model model) {
-		model.addAttribute("pmMems",service.getOnlyPeriod(date));
-		
-		return "/admin/pmMembers";
-	}
-	
 	// 차트 & 프리미엄 회원 조회 창으로 이동
 	@RequestMapping(value = "/profitChart", method = RequestMethod.GET)
 	public String profitChartGET(Model model) {
