+++ src/main/webapp/WEB-INF/views/board/readFood.jsp
@@ -3,7 +3,7 @@
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>    
 <%@ taglib prefix="fmt" uri="http://java.sun.com/jsp/jstl/fmt"%>    
 <jsp:include page="../include/header.jsp" />
-<link rel="stylesheet" href="${pageContext.request.contextPath}/resources/css/board.css" type="text/css">	
+<link rel="stylesheet" href="${pageContext.request.contextPath}/resources/css/board/board.css" type="text/css">
