+++ src/main/webapp/resources/css/board/board.css
@@ -2,6 +2,12 @@
 @import url("https://hangeul.pstatic.net/hangeul_static/css/nanum-square.css");
 @import url("https://hangeul.pstatic.net/hangeul_static/css/nanum-barun-gothic.css");
 
+#wrap {
+	width: 490px;
+	text-align: center;
+	margin: 0 auto 0 auto;
+}
+
 .ri-text {
 	margin-bottom: 3em;
 	border-bottom: 1px solid #E5E5E5;
@@ -45,8 +51,17 @@
 	font-size: 17px;
 }
 
+#modBtn_group button {
+	background-color: #6aafe6;
+	color: #d4dfe6;
+	border: 1px solid skyblue;
+	padding: 5px;
+	border-radius: 10px;
+	top: 400px;
+}
+
 #contentArea {
-	width: 100%;
+	width: 500px;
 	height: 200px;
 	padding: 10px;
 	box-sizing: border-box;
