--- src/main/webapp/resources/css/hyejin.css
@@ -1,348 +1,0 @@
-@charset "UTF-8";
-
-@import url("https://hangeul.pstatic.net/hangeul_static/css/nanum-square.css");
-@import url("https://hangeul.pstatic.net/hangeul_static/css/nanum-barun-gothic.css");
-
-	.area {
-	    position: absolute;
-	    background: #fff;
-	    border: 1px solid #888;
-	    border-radius: 3px;
-	    font-size: 18px;
-	    top: -5px;
-	    left: 15px;
-	    padding:2px;
-	}
-	
-	.info-menu {
-		font-size: 20px;
-		text-decoration: none !important;
-		color: #fff !important;
-	}
-	
-	
-	#load-more {
-		cursor: pointer;
-	
-	}
-	
-	
-	.ri-text {
-		margin-bottom: 3em;
-		border-bottom: 1px solid #e5e5e5;
-		padding-bottom: 2em;
-	}
-	
-	.ri-text h2 {
-		margin-bottom: 14px;
-		color: #01579B;
-	}
-	
-	.infoImage {
-		width: 360px;
-		height: 350px;
-		margin-right: 1em;
-	}
-	
-	.weatherPic {
-		width: 150px;
-		margin-right: 1em;
-	}
-	
-	.info {
-		margin-bottom: 3em;
-		border-bottom: 1px solid #e5e5e5;
-		padding-bottom: 2em;
-		font-family: 'NanumSquare' !important;
-	}
-	
-	.room-item .info {
-		border: 1px solid #ebebeb;
-		border-top: none;
-		padding: 24px 24px 30px 28px;
-		font-family: 'NanumSquare' !important;
-	}
-	
-	.room-item .info h4 {
-		color: #19191a;
-		margin-bottom: 17px;
-		font-family: 'NanumSquare' !important;
-	}
-	
-	.nanumFont {
-		font-family: 'NanumSquare' !important;
-	}
-	
-	#blogList * {
-		font-family: 'Noto Sans KR', sans-serif;
-	}
-	
-	#resInfo {
-		margin-top: 2em;
-	}
-	
-	#resInfo tr{
-		margin: 2em 0;
-	}
-	
-	
-	.table-wrapper {
-		-webkit-overflow-scrolling: touch;
-    	overflow-x: auto;
-	}
-	
-	.table-wrapper table {
-		margin: 0 0 2em 0;
-   		width: 100%;
-   		border-collapse: collapse;
-   	 	border-spacing: 0;
-	}
-	
-	.table-wrapper thead {
-		border-bottom: solid 2px rgba(210, 215, 217, 0.75);
-	}
-	
-	.table-wrapper th {
-		color: #3d4449;
-	    font-size: 1.1em;
-	    font-weight: 600;
-	    padding: 0 0.75em 0.75em 0.75em;
-	    text-align: left;
-	}
-	
-	.table-wrapper td {
-		padding: 0.75em 0.75em;
-	}
-	
-	.table-wrapper tbody tr:nth-child(2n + 1) {
-		background-color: rgba(230, 235, 237, 0.25);
-	}
-	
-	.table-wrapper tbody tr {
-		border: solid 1px rgba(210, 215, 217, 0.75);
-	    border-left: 0;
-	    border-right: 0;
-	}
-	
-	.table-wrapper tfoot {
-		border-top: solid 2px rgba(210, 215, 217, 0.75);
-	}
-	
-	.table-wrapper tfoot th {
-		padding: 1em 0.75em 0.75em 0.75em;
-	}
-	
-	#total {
-		font-style: italic;
-	}
-	
-	#pmSort {
-		margin-top: 1em;
-	} 
-	
-	.w-btn {
-	    position: relative;
-	    border: none;
-	    display: inline-block;
-	    padding: 10px 10px;
-	    border-radius: 15px;
-	    font-family: 'Noto Sans KR', sans-serif;
-	    text-decoration: none;
-	    font-weight: 600;
-	    transition: 0.25s;
-	    width: 130;
-	    height: 30;
-	    margin-right: 0.5em;
-	    
-	}
-	
-	.w-btn-indigo {
-	    background-color: aliceblue;
-	    color: #1e6b7b;
-	}
-	
-	
-	#profitBtn {
-	  -webkit-appearance: none;
-	  -moz-appearance: none;
-	  appearance: none;
-	  
-	  background: #0d6efd;
-	  color: #fff;
-	  
-	  margin: 0;
-	  padding: 0.3rem 1.3rem;
-	  
-	  font-family: 'Noto Sans KR', sans-serif;
-	  font-size: 0.9rem;
-	  font-weight: 400;
-	  text-align: center;
-	  text-decoration: none;
-	  
-	  border: none;
-	  border-radius: 4px;
-	  
-	  display: inline-block;
-	  width: auto;
-	  
-	  cursor: pointer;
-	  
-	  transition: 0.5s;
-	}
-
-
-	#profitBtn:active,
-	#profitBtn:hover,
-	#profitBtn:focus {
-	  background: #025ce2;
-	  outline: 0;
-	}
-	#profitBtn:disabled {
-	  opacity: 0.5;
-	}
-	
-	
-	 .periodBtn {
-	  -webkit-appearance: none;
-	  -moz-appearance: none;
-	  appearance: none;
-	  
-	  background: #fff;
-	  color: black;
-	  
-	  margin: -0.2em;
-	  padding: 0.2rem 1rem;
-	  
-	  font-family: 'Noto Sans KR', sans-serif;
-	  font-size: 0.9rem;
-	  font-weight: 400;
-	  text-align: center;
-	  text-decoration: none;
-	  
-	  display: inline-block;
-	  width: auto;
-	  
-	  cursor: pointer;
-	  
-	  transition: 0.5s;
-	  
-	  border: 1px solid #858796;
-	}
-
-
-	.periodBtn:active,
-	.periodBtn:hover,
-	.periodBtn:focus {
-	  background: #f0f0f5;
-	  outline: 0;
-	}
-
-	#like_img {
-		width: 30px;
-		height: 30px;
-		margin: 0;
-	}
-
-	button {
-		all: unset;
-	}
-	
-	
-	#keywordInput {
-		width: 470px;
-		border: none;
-		border-bottom: 1px solid black;
-	}
-	
-	.b-btn {
-    position: relative;
-    border: none;
-    display: inline-block;
-    padding: 10px 10px;
-    border-radius: 15px;
-    font-family: 'Noto Sans KR', sans-serif;
-    text-decoration: none;
-    font-weight: 600;
-    transition: 0.25s;
-    width: 130;
-    height: 30;
-    margin-right: 0.5em;
-	    
-	}
-	
-	.b-btn-indigo {
-	    background-color: black;
-	    color: #fff;
-	}
-	
-	#searchMenu {
-		transform: translate(0,80px);
-	}
-	
-	#searchAllMenu {
-		transform: translate(0,130px);
-	}
-	
-	#content-wrapper * {
-		font-family: 'Noto Sans KR', sans-serif !important;
-	}
-	
-	.one h2{
-	  position: relative;
-	  padding: 0;
-	  margin: 0;
-	  font-family: "Noto Sans KR", sans-serif;
-	  font-weight: 300;
-	  font-size: 40px;
-	  color: #080808;
-	  -webkit-transition: all 0.4s ease 0s;
-	  -o-transition: all 0.4s ease 0s;
-	  transition: all 0.4s ease 0s;
-	  padding-bottom: 5px;
-	  margin-left: 2.2em;
-	}
-	
-	.one h2:before {
-	  width: 28px;
-	  height: 5px;
-	  display: block;
-	  content: "";
-	  position: absolute;
-	  bottom: 3px;
-	  left: 50%;
-	  margin-left: -14px;
-	  background-color: #b80000;
-	}
-	.one h2:after {
-	  width: 100px;
-	  height: 1px;
-	  display: block;
-	  content: "";
-	  position: relative;
-	  margin-top: 25px;
-	  left: 50%;
-	  margin-left: -50px;
-	  background-color: #b80000;
-	}
-	
-	#mapTitle {
-		display: inline-block;
-		vertical-align: top;
-		width: 49%; 
-		margin-top: 13em;
-		text-align: center;
-	}
-	
-	#mapTitle p {
-		white-space: pre;
-		margin-left: -0.5em;
-	}
-	
-	#readMenu {
-		font-family: 'Noto Sans KR' !important;
-	}
-
-	#alignText a {
-		color: black;
-		font-family: 'Noto Sans KR', sans-serif;
-	}
