+++ src/main/webapp/resources/js/admin/vipInfo.js
@@ -11,9 +11,6 @@
 		return year+'-'+month+'-'+day;
 	}
 
-	$(function(){
-		
-		
 		$('#today').click(function(){
 			var date = new Date();
 			document.getElementById("startDate").value = formatDate(date);
@@ -49,8 +46,7 @@
 		});
 		
 		
-		
-		$('#profitBtn').click(function(){
+		function pmMems(){
 			var sDate = $("input[name=startDate]").val();
 			var eDate = $("input[name=endDate]").val();
 			var date = sDate+" "+eDate;
@@ -58,10 +54,90 @@
 			if(sDate == "" || eDate == ""){
 				alert("조회할 날짜를 선택해주세요");
 			} else {
-				location.href = '/admin/pmMembers?date='+date;
+				$('tbody').empty();
+				$('tfoot').empty();
+				$('#filterBtn').empty();
+					
+					let btnTag = '';
+					btnTag += "<input type='button' value='구독회원만 보기' class='w-btn w-btn-indigo' onclick='periodMems();'>";
+					btnTag += "<input type='button' value='모두보기' class='w-btn w-btn-indigo' onclick='pmMems();'>";
+				
+					$('#filterBtn').append(btnTag);
+					
+				$.ajax({
+					url : "/adminRest/pmMembers?date="+date,
+					success : function(data){
+					let tag = '';
+					if(data != ""){
+						$(data).each(function(i,obj){
+							tag += "<tr>";
+							tag += "<td>"+obj.id+"</td>";
+								 
+								 if(obj.pcd_PAYER_ID == null){
+									tag += "<td>N</td>";
+								 } else {
+									tag += "<td>Y</td>";
+								 }
+							tag += "<td>"+obj.create_date+"</td>";
+							tag += "<td>"+obj.license_deadline+"</td>";
+							tag += "</tr>";
+						});
+						$('tbody').append(tag);
+						$('tfoot').append("<tr><th colspan='3' id='total'>총 수익</th><th>"+($(data).length)*6000+"</th></tr>");
+						
+					} else {
+						$('tbody').append("<tr><td colspan='4'>내역이 없습니다.</td></tr>");
+						$('tfoot').append("<tr><th colspan='3' id='total'>총 수익</th><th>"+($(data).length)*6000+"</th></tr>");
+						}
+					}
+				}); // ajax
+			
+			}
+		
+		}
+		
+		
+		function periodMems(){
+			var sDate = $("input[name=startDate]").val();
+			var eDate = $("input[name=endDate]").val();
+			var date = sDate+" "+eDate;
+			
+			if(sDate == "" || eDate == ""){
+				alert("조회할 날짜를 선택해주세요");
+			} else {
+				$('tbody').empty();
+				$('tfoot').empty();
+				$('#filterBtn').empty();
+					
+					let btnTag = '';
+					btnTag += "<input type='button' value='구독회원만 보기' class='w-btn w-btn-indigo' onclick='periodMems();'>";
+					btnTag += "<input type='button' value='모두보기' class='w-btn w-btn-indigo' onclick='pmMems();'>";
+				
+					$('#filterBtn').append(btnTag);
+					
+				$.ajax({
+					url : "/adminRest/periodMems?date="+date,
+					success : function(data){
+					let tag = '';
+					if(data != ""){
+						$(data).each(function(i,obj){
+							tag += "<tr>";
+							tag += "<td>"+obj.id+"</td>";
+							tag += "<td>Y</td>";
+							tag += "<td>"+obj.create_date+"</td>";
+							tag += "<td>"+obj.license_deadline+"</td>";
+							tag += "</tr>";
+						});
+						$('tbody').append(tag);
+						$('tfoot').append("<tr><th colspan='3' id='total'>총 수익</th><th>"+($(data).length)*6000+"</th></tr>");
+						
+					} else {
+						$('tbody').append("<tr><td colspan='4'>내역이 없습니다.</td></tr>");
+						}
+					}
+				}); // ajax
+			
 			}
 			
-			
-		});
+		}
 		
-	}); // JQuery
